pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-mpm-agents"
version = "0.1.0"
description = "Claude MPM Agents with DeepEval testing"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["tests*"]

[project.optional-dependencies]
test = [
    "deepeval>=2.5.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pydantic>=2.0.0",
    "pytest-cov>=4.1.0",
]
dev = [
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "instruction_compliance: Tests for BASE-AGENT.md instruction compliance",
    "role_boundary: Tests for agent role boundary enforcement",
    "output_quality: Tests for response quality metrics",
    "registry: Tests for agent registry validation",
]

[tool.ruff]
line-length = 100
target-version = "py310"
